// ordermanager-ui/ui/resources/frontend/invoices/src/app/workflows/invoice-workflow/state/invoice.state.ts
export enum WorkflowStatuses {
  SET_INVOICE_TYPE = 'SET_INVOICE_TYPE',
  SET_INVOICE_DATE = 'SET_INVOICE_DATE',
  SET_INVOICE_CREATOR = 'SET_INVOICE_CREATOR',
  SET_INVOICE_RECIPIENT = 'SET_INVOICE_RECIPIENT',
  SET_INVOICE_ITEMS = 'SET_INVOICE_ITEMS',
  SAVE_INVOICE = 'SAVE_INVOICE',
  SAVE_INVOICE_ITEM = 'SAVE_INVOICE_ITEM'
}